feat(custom-path): extraPageFields at page level
This commit adds the `extraPageFields` key to the `Page` interface.  Previously it was only on the `PluginOptions` interface, which is overly restrictive because it requires every page type to have the same fields, lest graphql throw an error that the page doesn't exist.

Although I don't think that this field belongs on the global plugin options, I left it there and handled it for backwards compatibility.